Use PromQL to widen the data horizon
PromQL seems to be able to get more results than the log cache API.
However it is not a perfect fit for out use case and it has an annoying
bug - it does not interpret regex patterns, bit uses them as plain
values. We worked around this by further filtering the results in the
plugin.

Co-authored-by: Giuseppe Capizzi <gcapizzi@pivotal.io>