Ignore metrics for instances that do not exist
Scaling down CF apps results into destroying instances. Therefore if we
get metrics for instances that have been destroyed, we ignore them.

Also, when the CF app is scaled down and then back up we would get two
series of metrics with the same instance id: one for the current
instance and one for the one that got destroyed during scale down. In
order to figure out which is the series for the current instance, we
check the series first point timestamp to be greater than the instance
creation time.

[finishes #166284175]

Co-authored-by: Giuseppe Capizzi <gcapizzi@pivotal.io>