Use cf client's choice of SSL options
If CF_API is http://, use http:// for log-cache.
If it is https://, use https:// for log-cache.
If --skip-ssl-validation is active, also skip for log-cache.

If valid certificates are in the cert pool on the machine, these will be
used. Also, on linux, SSL_CERT_FILE and SSL_CERT_DIR will be respected
by the go tls library.

[#168075720]